+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+- Added: `PageError` component can be used in `PageContainer` when page have
+errors.
+- Added: `PageLoading` component can be used in `PageContainer` when page are
+loading.
+- Removed: `NotFound` and `Loading` props of `PageContainer`
 
 # 0.3.0 - 2015-10-30